requirements.txt
fastapi
uvicorn
tensorflow
keras
numpy
pillow
streamlit
requests
